Add hora de creación y verificar que se creo correctamente el archivo rar

+++ projects/gmm-emision-uat.bash
@@ -10,17 +10,28 @@ createRar() {
     anio=$(date -d "$fecha" "+%Y")
     mes=$(date -d "$fecha" "+%m")
     dia=$(date -d "$fecha" "+%d")
+    hora=$(date -d "$fecha" "+%H")
+    minuto=$(date -d "$fecha" "+%M")
+    segundo=$(date -d "$fecha" "+%S")
 
     # Crear el nombre del archivo rar
-    nombre_archivo="gmm-emision@$anio-$mes-$dia.rar"
+    nombre_archivo="gmm-emision@$anio-$mes-$dia-$hora-$minuto-$segundo.rar"
+
+    echo "Creando archivo rar: $nombre_archivo"
 
     # Crear el archivo rar
     rar a "$nombre_archivo" "$ruta/index.html" "$ruta/vite.svg" "$ruta/assets"
 
-    # Eliminar los archivos y carpetas
-    rm "$ruta/index.html"
-    rm "$ruta/vite.svg"
-    rm -r "$ruta/assets"
+    # Verificar que se creó el archivo rar correctamente
+    if [ -f "$nombre_archivo" ] && [ $(stat -c "%s" "$nombre_archivo") -gt 0 ]; then
+        # Eliminar los archivos y carpetas
+        rm "$ruta/index.html"
+        rm "$ruta/vite.svg"
+        rm -r "$ruta/assets"
+        echo "El archivo rar se creó correctamente: $nombre_archivo"
+    else
+        echo "Error: no se creó el archivo rar"
+    fi
 }
 
 # Función para publicar en un entorno específico
